Initial upload of Visual Studio 2013 solution and project files.
This contains on project that builds the pbrt code base as a static library (although I am getting compile errors with the code itself), which will be used by the various executable projects.  It also has a shared project properties file that adds the include search paths to any project that inherits it.