--------------------------
Made stdafx.h a no-op.

pbrt.h put up the include guard for the body of the file, then included stdafx.h, which then included pbrt.h again, but by that point PBRT_CORE_PBRT_H is already defined, so the body of pbrt.h was locked out, causing problems with headers in stdafx.h which depend on pbrt.h.

--------------------------

Added work arounds for a couple of built in GCC functions for MSVC,  One of these was already in the comments.  ** Need to check MSVC implementation of __builtin_clz().  **

--------------------------

VS2013 does not support "constexpr" keyword, so some constants use define instead on MSVC as a workaround.

--------------------------