Add --toply command line option
Similar to --cat, in that it prints an equivalent scene file to stdout, but it also creates a PLY file for each triangle mesh in the scene and rewrites the scene file to use them.

Removed the PBRT_DUMP_PLY environment variable hack to dump a ply file for each triangle mesh, which was less useful.