Ensure samples are in [0,1) after remapping.
In BSDF::Sample_f(), u[0] is first used to determine which BSDF component to sample; it is then remapped to [0,1) for use in the actual sampling routine. Before, due to floating-point roundoff error, it was possible that it would take on the value 1, which in turn led to NaN values in the Trowbridge-Reitz sampling routines.

Fixed the same issue in FresnelBlend::Sample_f() just to be safe.