Be careful to not try to normalize a degenerate vector in BDPT.
It turns out that with participating media, the MLT integrator in
particular sometimes creates paths with two vertices at the same position,
which in turn was leading to trying to normalize degenerate vectors in the
BDPT code, which in turn led to NaN values. (This happened in particular
with the volume-caustic scene.)

The specific cause was sample values of 0 from the MLTSampler, which in
turn lead to sampled t values of 0 in HomogeneousMedium::Sample(). (After a
bunch of investigation, it became clear that these 0 values were
legitimate, just things randomly lining up so that a small step mutation
would end up exactly at 0.)