Fix profiler segfault fix.
Actually, using a unique_ptr at all is a bad idea (with multiple
threads). The issue is that even if we reset the pointer at shutdown time,
there may be a race where another thread has read its value and is
modifying its memory while the main thread is freeing it.

Therefore, allocate profileSamples at startup time, but don't worry about
freeing it. This is just a one-time memory leak (and not growing), so isn't
really a problem.

Issue #83.