Clean up handling of per-thread light distribution caches.
Rather than using a global thread local variable, keep a vector in
the main SpatialLightDistribution class that is accessed using
ThreadIndex. Not only does this clean up the implementation (and
allow multiple SpatialLightDistributions to exist), it seems to clear
up intermittent crashes due to freeing bad memory in the
SpatialLightDistribution destructor.