Fix rare infinite loop bug in SampleFourier().
In very rare circumstances, the while(true) loop in that function would
continue forever. What was happening was that we'd get stuck on a
particular bracketing interval [a,b], wherein loop iterations would
alternately just so happen to compute phi=a and then phi=b.  In turn, no
progress would be made and the loop would continue forever.

This is easily fixed by choosing the bisection step when phi exactly equals
one of the endpoints.

(Though this bug hasn't yet been seen in SampleCatmullRom() or
InvertCatmullRom(), there is similar logic in those functions, so also
updated the logic there similarly.)