Fix bug in SeparableBSSRDF::Sample_Sp().
In rare cases, it would compute a zero value for the |l| variable,
which in turn led to pTarget being equal to the intersection point,
which in turn led to a ray with a (0,0,0) direction vector, which
in turn led to NaN values being computed in Triangle::Intersect().
(It’s surprising that it got to Triangle::Intersect() in the first
place with such a ray.)

Fixes issue #93.