Fix bug in Sphere::Sample().
Because the sampling coordinate system is in world space, the surface
normal computed by SphericalDirection() is itself already in world
space. However, we were treating it as being in object space, which in turn
led to errors if the sphere's object-to-world transformation involved
anything more than translations.

The following scene illustrates the bug: the rotation applied to the sphere
should have no effect on the final image, since the sphere starts out at
the origin, but without this bug fix, the image changes substantially if
it's there.

WorldBegin

AttributeBegin
  Translate 0 2 0
  Rotate 180 1 0 0
  AreaLightSource "diffuse" "color L" [40 40 40]
  Shape "sphere" "float radius" .5
AttributeEnd

Material "matte"
Shape "trianglemesh" "point P" [-5 -1 0  5 -1 0   0 -1 5]
  "integer indices" [0 1 2]

WorldEnd