Fix normal calculation in Triangle::Sample().
We used to interpolate a shading normal and return that, which doesn't seem
like the right thing to do. As in intersect, we do use the shading normal
to orient the geometric normal, if shading normals are present.

Also made the handling of 'reverse orientation' and 'transform swaps
handedness' consistent with elsewhere: only flip the normal for these if
there is no shading normal (and check the transform handedness stuff, which
wasn't being done before.)

Checked a number of pbrt-v3-scenes: no meaningful difference in san miguel,
measure one, contemporary bathroom, barcelona pavillion (night), or crown,
all of which have triangle area lights with per-vertex surface normals.
bathroom does change, due to the expectation that 'reverse orientation'
applied even if there were per-vertex normals (now no longer the case). A
fix for this will be pushed to the scenes distribution.