Fix handling of per-vertex normals with transforms that swap handedness
With triangle meshes, even when the shading normal's direction was authoratative (coming via per-vertex normal vectors), we'd still flip the shading normal if the transform swapped the coordinate system handedness. This in turn led to bugs where the wrong side of the light would be emissive when a ray intersected it.

Issue #229